+++ app.py
@@ -7,7 +7,7 @@ from decouple import config as en_var # import the environment var
 # import pytz
 
 print("SystemInfo -> ", systemInfo)
-print("Environment Variable: "+ en_var('christmas_app2022'))
+# print("Environment Variable: "+ en_var('christmas_app2022'))
 
 # print(pytz.all_timezones) # List out all the timezone available
 if __name__ == '__main__':